+++ packages/react-native-messaging-core/android/src/main/java/com/nabla/sdk/reactnative/messaging/core/nablamessagingclient/ConversationItemsWatcherModule.kt
@@ -111,8 +111,8 @@ internal class ConversationItemsWatcherModule(
         if (count == 0) {
             conversationItemsWatchersJobs.forEach {
                 it.value.first.cancel()
-                conversationItemsWatchersJobs.remove(it.key)
             }
+            conversationItemsWatchersJobs.clear()
         }
     }