+++ packages/react-native-messaging-core/src/NablaMessagingClient/NativeModules/index.ts
@@ -1,0 +1,4 @@
+export * from './ConversationItemsWatcherModule';
+export * from './ConversationListWatcherModule';
+export * from './ConversationWatcherModule';
+export * from './NablaMessagingClientModule';