+++ packages/react-native-messaging-core/src/NablaMessagingClient/types/index.ts
@@ -1,0 +1,6 @@
+export * from './errorMapper';
+export * from './NativeConversation';
+export * from './NativeConversationItem';
+export * from './NativeConversationItems';
+export * from './NativeConversationList';
+export * from './NativeProviderInConversation';