+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `NablaMessagingClient.watchItemsOfConversation` to watch conversation items with a given `conversationId`.
+- `NablaMessagingClient.watchConversation` to watch a conversation with a given id.
 
 ### Changed